requirements.txt
fastapi==0.85.0
pydantic==1.10.2
uvicorn==0.18.3
peewee==3.15.3
cryptography==38.0.3
peewee-migrate==1.6.2
python-dotenv==0.21.0
pymysql==1.0.2
python-multipart==0.0.5
python-slugify==6.1.2
